requirements.txt
requests>=2.28.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
pyrogram>=2.0.106,<3
TgCrypto>=1.2.5
httpx>=0.27.0
